+++ src/pptx/oxml/__init__.py
@@ -448,6 +448,9 @@ from pptx.oxml.text import (  # noqa: E402
     CT_TextAutonumberBullet,
     CT_TextBody,
     CT_TextBodyProperties,
+    CT_TextBulletColor,
+    CT_TextBulletSizePercent,
+    CT_TextBulletSizePoint,
     CT_TextCharBullet,
     CT_TextCharacterProperties,
     CT_TextField,
@@ -466,7 +469,11 @@ register_element_cls("a:bodyPr", CT_TextBodyProperties)
 register_element_cls("a:br", CT_TextLineBreak)
 register_element_cls("a:buAutoNum", CT_TextAutonumberBullet)
 register_element_cls("a:buChar", CT_TextCharBullet)
+register_element_cls("a:buClr", CT_TextBulletColor)
+register_element_cls("a:buFont", CT_TextFont)
 register_element_cls("a:buNone", CT_TextNoBullet)
+register_element_cls("a:buSzPct", CT_TextBulletSizePercent)
+register_element_cls("a:buSzPts", CT_TextBulletSizePoint)
 register_element_cls("a:defRPr", CT_TextCharacterProperties)
 register_element_cls("a:endParaRPr", CT_TextCharacterProperties)
 register_element_cls("a:fld", CT_TextField)